% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCarb-package.R
\docType{data}
\name{Example_Data}
\alias{Example_Data}
\title{Example data}
\format{\code{Example_Data}: \link{data.frame}}
\description{
Example data
}
\section{Version}{
 0.1.0
}

\examples{

## show first 5 elements of the METADATA and DATA elements in the terminal
data(Example_Data, envir = environment())
head(Example_Data)

}
\keyword{datasets}
