% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_InputTemplate.R
\name{write_InputTemplate}
\alias{write_InputTemplate}
\title{Write table input template}
\usage{
write_InputTemplate(file = NULL, nrows = NULL, ...)
}
\arguments{
\item{file}{\link{character} (optional): output path, if \code{NULL} nothing is written, but a template
\link{data.frame} is returned.}

\item{nrows}{\link{integer} (optionl): set the number of rows in the template, the default \code{NULL} is one.}

\item{...}{additional arguments that can be passed to function \link{write.table} if \code{file != NULL}.
Supported arguments are: \code{sep}, \code{dec}, \code{fileEncoding}}
}
\description{
This function creates a template table that can be used as input for the function
\link{model_DoseRate}
}
\section{Function version}{
  0.1.0
}

\examples{

##create template without file creation
write_InputTemplate()

\dontrun{
##Example with file output

## set temporary filename
## (replace by own path if needed)
temp_file <- tempfile(pattern = "template", fileext = ".csv")
write_InputTemplate(file = temp_file)

}

}
\seealso{
\link{Example_Data}, \link{write.table}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
