% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DoseRate.R
\name{model_DoseRate}
\alias{model_DoseRate}
\title{Model dose rate evolution in carbonate-rich samples}
\usage{
model_DoseRate(data, length_step = 1L, max_time = 500L, n.MC = 100,
  method_control = list(), txtProgressBar = TRUE, verbose = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{data}{\link{data.frame} (\strong{required}): input data following the structure given
in the example data set \code{data(Example_Data)}. The input \link{data.frame} should have at least
one row (i.e. values for one sample). For multiple rows the function is automatically re-called.}

\item{length_step}{\link{numeric} (with default): step length used for the calculation}

\item{max_time}{\link{numeric} (with default): maximum temporal search range}

\item{n.MC}{\link{numeric} (with default): number of Monte Carlo runs used for the error calculation}

\item{method_control}{(\emph{optional}): additional arguments that can be provided to the control the
the modelling. See details for further information.}

\item{txtProgressBar}{\link{logical} (with default): enables/disables the \code{txtProgressBar} for the MC runs}

\item{verbose}{\link{logical} (with default): enables/disables verbose mode}

\item{plot}{\link{logical} (with default): enables/disables plot output}

\item{...}{further arguments passed to the underyling plot functions, see also details for further information. Supported standard arguments are \code{mfrow}, \code{xlim}, \code{xlab}.}
}
\value{
The function returns numerical and graphical output

-----------------------------------\cr
\code{[ NUMERICAL OUTPUT ]}\cr
-----------------------------------\cr
\itemize{
\item A \link{data.frame} which is the combination of the input and values calculated by this function.
}

-----------------------------------\cr
\code{[ GRAPHICAL OUTPUT ]}\cr
-----------------------------------\cr
\itemize{
\item Two plots are returned: ##TODO
}
}
\description{
This function models the dose rate evolution in carbonate enrich environments. For the
calculation internal functions are called.
}
\details{
TODO
}
\section{Function version}{
 0.1.0
}

\examples{
##load example data
data("Example_Data", envir = environment())

##run the function for one sample from
##the dataset
model_DoseRate(
data = Example_Data[14,],
n.MC = 2,
txtProgressBar = FALSE
)


}
\references{
Mauz, B., Hoffmann, D., 2014. What to do when carbonate replaced water: Carb, the model for estimating the
dose rate of carbonate-rich samples. Ancient TL 32, 24–32.

Nathan, R.P., Mauz, B., 2008. On the dose-rate estimate of carbonate-rich sediments for trapped charge dating.
Radiation Measurements 43, 14–25. doi:10.1016/j.radmeas.2007.12.012 \cr

\strong{Further reading}

Nathan, R.P., 2010. Numerical modelling of environmental dose rate and its application to trapped-charge dating.
DPhil thesis, St Hugh’s College, Oxford.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, Université Bordeaux Montagine (France); based
on MATLAB code given in Carb_2007a
}
