% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_InputTemplate.R
\name{write_InputTemplate}
\alias{write_InputTemplate}
\title{Write table input template}
\usage{
write_InputTemplate(file = NULL, ...)
}
\arguments{
\item{file}{\link{character} (optional): output path, if \code{NULL} nothing is written, but a template
\link{data.frame} is returned.}

\item{...}{additional arguments that can be passed to function \link{write.table} if \code{file != NULL}.
Supported arguments are: \code{sep}, \code{dec, }fileEncoding`}
}
\description{
This function creates a template table that can be used as input for the function
\link{model_DoseRate}
}
\section{Function version}{
  0.1.0
}

\examples{

##create template without file creation
write_InputTemplate()

\dontrun{
##Example with file output
##To run: uncomment and modify path
#write_InputTemplate(file = "[YOUR PATH]")

}

}
\seealso{
\link{Example_Data}, \link{write.table}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
}
