% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_InputTemplate.R
\name{write_InputTemplate}
\alias{write_InputTemplate}
\title{Write table input template}
\usage{
write_InputTemplate(file = NULL, ...)
}
\arguments{
\item{file}{\link{character} (optional): output path, if \code{NULL} nothing is written, but a template
\link{data.frame} is returned.}

\item{...}{additional arguments that can be passed to function \link{write.table} if \code{file != NULL}.
Supported arguments are: \code{sep}, \code{dec, }fileEncoding`}
}
\description{
This function creates a template table that can be used as input for the function
\link{model_DoseRate}
}
\section{Function version}{
  0.1.0
}

\examples{

##create template without file creation
write_InputTemplate()

\dontrun{
##Example with file output

## set temporary filename
## (replace by own path if needed)
temp_file <- tempfile(pattern = "template", fileext = ".csv")
write_InputTemplate(file = temp_file)

}

}
\seealso{
\link{Example_Data}, \link{write.table}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
} 

\section{How to cite}{
Kreutzer, S., 2019. write_InputTemplate(): Write table input template. Function version 0.1.0. In: Kreutzer, S., Nathan, R.P., Mauz, B., 2019. RCarb: Dose Rate Modelling of Carbonate-Rich Samples R package version 0.1.3.9000-21. 
}

